requirements.txt
flask
python-dotenv
azure-ai-translation-text
azure-storage-blob
azure-ai-translation
